pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/jrhrsmit/ppk-pd/issues"


[tool.poetry]
name = "ppk_pd"
version = "0.0.1"
authors = ["Jasper Smit"]
readme = ["README.md", "LICENSE"]
description = "Power profiler with USB power delivery and current/voltage limits"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
repository = "https://github.com/jrhrsmit/ppk-pd/tree/feature/base"
homepage = "https://github.com/jrhrsmit/ppk-pd/tree/feature/base"

[tool.poetry.dependencies]
python = "^3.11,<3.13"                               # max allowed version by scipy
faebryk = { path = "../../faebryk", develop = true }
typer = { version = "^0.9.0", extras = ["all"] }

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pytest = "^7.1.3"
isort = "^5.6.4"
ruff = "^0.0.291"
wget = "^3.2"
tortoise-orm = "^0.20.0"
si-prefix = "^1.2.2"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
    # Pycodestyle
    "E",
    "W",
    # Pyflakes
    "F",
    # Isort
    "I001",
]
ignore = [
    # Line too long
    #"E501",
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = [
    "A",
    "B",
    "C",
    "D",
    "E",
    "F",
    "G",
    "I",
    "N",
    "Q",
    "S",
    "T",
    "W",
    "ANN",
    "ARG",
    "BLE",
    "COM",
    "DJ",
    "DTZ",
    "EM",
    "ERA",
    "EXE",
    "FBT",
    "ICN",
    "INP",
    "ISC",
    "NPY",
    "PD",
    "PGH",
    "PIE",
    "PL",
    "PT",
    "PTH",
    "PYI",
    "RET",
    "RSE",
    "RUF",
    "SIM",
    "SLF",
    "TCH",
    "TID",
    "TRY",
    "UP",
    "YTT",
]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"

[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10
